break out I/O checks to common testing, move streaming tests
